Major change in crowdsourcing support, and others
1. Added a braincatalogue font, all icons should go here in the future
2. modifications of css for responsive design: specimen pages should
resize gracefully up to iphone size. An intermediate size layout would
be required: an issue was added for this
3. AtlasMaker is now structured as a widget, and was added to the
specimen pages. I.e., instead of separate pages for specimen info and
atlas edition, everything happens now in the same page (and then, the
MRI file has to be loaded only once for the two things)
4. atlasmakerserver was also updated. During the same websocket
connection a user can now view/edit many atlases. The code now handles
this
5. There’s is the issue of the long time required to load very large
MRI files. Originally, 1 sagittal, 1 axial and 1 coronal jpg pictures
were displayed during the time the MRI is being loaded. Now, the
server-side drawNiiSlice generates jpg files on the fly to display not
only the mid sag, axi and cor slices, but any slice the user would
require. This is a first step towards a fastest data loading
6. myMesh should also be widgetised. The new code includes changes in
this direction. An issue was added for this.
7. myStereotaxic.js was deleted — it’s no longer used. Now atlasMaker
does viewing and editing
8. the node module jpeg-js is used by drawNiiSlice to draw jpegs on the
fly
9. three.js and friends was updated to the latest version
10. braincatalogue.php now handles the generation of slice jpegs on the
fly (it calls drawNiiSlice)